Move back CDK TS file output location
